Edit trip in modal
