+++ frontend/package.json
@@ -11,6 +11,7 @@
     "moment": "^2.25.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-responsive-modal": "^5.0.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
     "styled-components": "4.2.0"
