+++ frontend/src/Assets/styles.css
@@ -1,0 +1,4 @@
+.customModal {
+    max-width: 600px;
+    width: 100%;
+}
