+++ frontend/src/routes.js
@@ -17,7 +17,6 @@ import TripListPastView from './Containers/TripListPastView';
 import TripListFutureView from './Containers/TripListFutureView';
 import TripListCancelledView from './Containers/TripListCancelledView';
 import TripCreateView from './Containers/TripCreateView';
-import TripEditView from './Containers/TripEditView';
 import FlightCreateView from './Containers/FlightCreateView';
 
 const BaseRouter = () => (
@@ -28,7 +27,6 @@ const BaseRouter = () => (
                 <Route exact path='/profile' component={ProfileView} />
                 <Route exact path='/trips' component={TripListView} />
                 <Route exact path='/trips/:tripID' component={TripDetailView} />
-                <Route exact path='/trips/edit/:tripID' component={TripEditView} />
                 <Route exact path='/past-trips' component={TripListPastView} />
                 <Route exact path='/active-trips' component={TripListFutureView} />
                 <Route exact path='/cancelled-trips' component={TripListCancelledView} />
